Refactor and combine ir test cases
