+++ test/wat_compiler/integration.test.ts
@@ -1,6 +1,6 @@
 import { compile, parse } from '../src';
-import { TEST_EXPORTS, encodeModule } from '../src/wat2wasm/binary_writer';
-import { getIR } from '../src/wat2wasm/ir';
+import { TEST_EXPORTS, encodeModule } from '../src/wat_compiler/binary_writer';
+import { getIR } from '../src/wat_compiler/ir';
 import { invalidTestCases as invalidFuncExpTestCases } from './resources/function_expressions';
 import {
   type ModuleTestCase,
