+++ test/wat_compiler/resources/invalid_function_bodies.ts
